+++ server/html-requests.js
@@ -1,0 +1,43 @@
+'use strict';
+
+function getRequest(url) {
+  return new Promise((resolve) => {
+    var xhr;
+    if (window.XMLHttpRequest) {
+      xhr = new XMLHttpRequest();
+    } else if (window.ActiveXObject) {
+      xhr = new ActiveXObject('Microsoft.XMLHTTP');
+    }
+    xhr.onreadystatechange = () => {
+      if (xhr.readyState === 4) {
+        resolve(xhr.responseText);
+      }
+    };
+    xhr.open('GET', url, true);
+    xhr.send();
+  });
+}
+
+function postRequest(url, data) {
+  return new Promise((resolve) => {
+    let xhr;
+    if (window.XMLHttpRequest) {
+      xhr = new XMLHttpRequest();
+    } else {
+      xhr = new ActiveXObject('Microsoft.XMLHTTP');
+    }
+    xhr.onreadystatechange = () => {
+      if (xhr.readyState === 4) {
+        resolve(xhr.responseText);
+      }
+    }
+    xhr.open('POST', url, true);
+    xhr.setRequestHeader('Content-Type', 'application/json');
+    xhr.send(JSON.stringify(data));
+  });
+}
+
+export {
+  getRequest as makeGetRequest,
+  postRequest as makePostRequest
+}
