+++ server/package.json
@@ -2,13 +2,17 @@
   "dependencies": {
     "date-fns": "^2.25.0",
     "express": "^4.17.1",
-    "node-static": "^0.7.11"
+    "jasmine": "^3.10.0",
+    "node-static": "^0.7.11",
+    "webpack": "^5.58.2",
+    "webpack-cli": "^4.9.0"
   },
   "name": "server",
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jasmine",
+    "build": "webpack"
   },
   "author": "",
   "license": "ISC",
