+++ server/webpack.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  mode: 'production',
+  watch: true,
+  entry: './script.js',
+  output: {
+    filename: './build.js',
+  },
+};
